Adds V Programming Language to the Fibonacci benchmarks. Fixes and issue with Numba on the json_loads benchmark. Modifies Nim code to use uint64 for Fibonacci benchmarks so as to not have the overflow exception.

+++ fibonacci_iterative/_Numba/fib_iter.py
@@ -1,6 +1,6 @@
-from numba import jit, int32, int64
+from numba import jit, int64
 
-@jit(int64(int32))
+@jit(int64(int64))
 def fibIter(n):
     a, b = 0, 1
     while n > 0:
